// lib/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      profiles: {
        Row: {
          company_id: string | null
          company_name: string | null
          created_at: string | null
          email: string | null
          full_name: string | null
          id: string
          phone: string | null
          role: 'super_admin' | 'owner' | 'driver' | null
          updated_at: string | null
          username: string | null
        }
        Insert: {
          company_id?: string | null
          company_name?: string | null
          created_at?: string | null
          email?: string | null
          full_name?: string | null
          id: string
          phone?: string | null
          role?: 'super_admin' | 'owner' | 'driver' | null
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          company_id?: string | null
          company_name?: string | null
          created_at?: string | null
          email?: string | null
          full_name?: string | null
          id?: string
          phone?: string | null
          role?: 'super_admin' | 'owner' | 'driver' | null
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}